Add tests
* add tests for layer

* restructure test data to mimic actual data structure

* update conftest to reflect new test data structure
